+++ packages/mesh/lists/index.ts
@@ -13,6 +13,9 @@ import InvitationCode from './invitation_code'
 import Image from './image'
 import Notify from './notify'
 import Announcement from './announcement'
+import Policy from './policy'
+import Transaction from './transaction'
+import Sponsorship from './sponsorship'
 
 export const listDefinition = {
   User,
@@ -30,4 +33,7 @@ export const listDefinition = {
   Notify,
   Announcement,
   Photo: Image,
+  Policy,
+  Transaction,
+  Sponsorship,
 }