+++ packages/mesh/lists/publisher.ts
@@ -30,7 +30,12 @@ const listConfigurations = list ({
 	  validation: {
         isRequired: true,
       },
-	  isindexed: 'unique',
+	    isindexed: 'unique',
+    }),
+    sponsored: relationship({
+      label: 'Sponsored',
+      ref: 'Sponsorship.publisher',
+      many: true,
     }),
 	lang: select({
 	  label: '語系',
@@ -63,9 +68,10 @@ const listConfigurations = list ({
 		{ label: '鏡週刊', value: 'mirrormedia' }, 
 		{ label: 'READr', value: 'readr' },
 		{ label: '鏡週刊 external', value: 'mm_external' },
-		{ label: 'rss', value: 'rss' }
+		{ label: 'rss', value: 'rss' },
+    { label: 'empty', value: 'empty'}
       ],
-	  defaultValue: 'rss',
+	  defaultValue: 'empty',
 	}),
     paywall: checkbox({
       defaultValue: false,