+++ packages/mesh/migrations/20240708034400_add_category_priority/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Category" ADD COLUMN     "priority" INTEGER NOT NULL DEFAULT 0;