+++ packages/mesh/migrations/20240722053122_add_member_wallet/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Member" ADD COLUMN     "wallet" TEXT NOT NULL DEFAULT E'';