+++ packages/mirrordaily/migrations/20240606050815_add_section_color/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Section" ADD COLUMN     "color" TEXT NOT NULL DEFAULT '#FFFFFF';